embedding-sandbox/requirements-frankenstein.txt
# Frankenstein Embeddings Requirements
# Updated for huggingface_hub compatibility
sentence-transformers>=3.0.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
psycopg2-binary>=2.9.0
